Add files via upload
This software is a basic Client Management System named as ClientSol. All the compiled code is in this package with external .jar files used.
To install this software, you first need to create its .jar file (can use Launch4j software), after that create a setup wizard for this software using  InnoDB compiler. This will create a .exe file for this software, which is a distributable version of this software.